# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple service to upload and retrieve PDFs.",
        "title": "PDF Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/extract/{id}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Extract text from PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PDF ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/pdf/{id}": {
            "get": {
                "description": "Returns the raw PDF file",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Retrieve a PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PDF ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Search PDFs by text",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Upload, store, extract text, and index",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Upload a PDF file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "PDF file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID of uploaded PDF",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}